Conflicts:
	lib/fog/aws/requests/storage/get_object.rb
	tests/helper.rb

+++ lib/fog/storage.rb
@@ -8,6 +8,9 @@ module Fog
     def self.new(attributes)
       attributes = attributes.dup # prevent delete from having side effects
       case provider = attributes.delete(:provider).to_s.downcase.to_sym
+      when :atmos
+        require 'fog/atmos/storage'
+        Fog::Storage::Atmos.new(attributes)
       when :aws
         require 'fog/aws/storage'
         Fog::Storage::AWS.new(attributes)
@@ -20,6 +23,9 @@ module Fog
       when :ibm
         require 'fog/ibm/storage'
         Fog::Storage::IBM.new(attributes)
+      when :internetarchive
+        require 'fog/internet_archive/storage'
+        Fog::Storage::InternetArchive.new(attributes)
       when :local
         require 'fog/local/storage'
         Fog::Storage::Local.new(attributes)
@@ -29,6 +35,9 @@ module Fog
       when :rackspace
         require 'fog/rackspace/storage'
         Fog::Storage::Rackspace.new(attributes)
+      when :openstack
+        require 'fog/openstack/storage'
+        Fog::Storage::OpenStack.new(attributes)
       else
         raise ArgumentError.new("#{provider} is not a recognized storage provider")
       end
